# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gfolio Admin & Corporate Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: kycStatus
          in: query
          schema:
            type: string
            enum: [pending, verified, rejected, all]
        - name: search
          in: query
          schema:
            type: string
        - name: userType
          in: query
          schema:
            type: string
            enum: [individual, corporate, all]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}/kyc:
    patch:
      operationId: updateUserKyc
      tags: [users]
      summary: Update user KYC status
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycUpdateRequest"
      responses:
        "200":
          description: KYC updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        kycStatus:
          type: string
          enum: [pending, verified, rejected]
        userType:
          type: string
          enum: [individual, corporate]
        totalInvested:
          type: number
        joinedAt:
          type: string
        lastActive:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - kycStatus
        - userType
        - totalInvested
        - joinedAt
        - lastActive
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        kycStatus:
          type: string
          enum: [pending, verified, rejected]
        kycDocuments:
          type: array
          items:
            type: string
        userType:
          type: string
          enum: [individual, corporate]
        totalInvested:
          type: number
        portfolioValue:
          type: number
        giftsReceived:
          type: integer
        giftsSent:
          type: integer
        joinedAt:
          type: string
        lastActive:
          type: string
        address:
          type: string
        panNumber:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - kycStatus
        - userType
        - totalInvested
        - portfolioValue
        - giftsReceived
        - giftsSent
        - joinedAt
        - lastActive
    KycUpdateRequest:
      type: object
      properties:
        status:
          type: string
          enum: [verified, rejected]
        remarks:
          type: string
      required:
        - status
...