# lib/api-zod/src/generated/types/kycUpdateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gfolio Admin & Corporate Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { KycUpdateRequestStatus } from "./kycUpdateRequestStatus";

export interface KycUpdateRequest {
  status: KycUpdateRequestStatus;
  remarks?: string;
}