# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gfolio Admin & Corporate Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UserKycStatus } from "./userKycStatus";
import type { UserUserType } from "./userUserType";

export interface User {
  id: string;
  name: string;
  email: string;
  phone: string;
  kycStatus: UserKycStatus;
  userType: UserUserType;
  totalInvested: number;
  joinedAt: string;
  lastActive: string;
}